+++ src/cupcnn/optimizer/SGDOptimizer.java
@@ -11,12 +11,12 @@ import cupcnn.data.Blob;
 
 public class SGDOptimizer extends Optimizer {
 	
-	public SGDOptimizer(double lr){
+	public SGDOptimizer(float lr){
 		super(lr);
 	}
 
 	
-	public SGDOptimizer(double lr,Optimizer.GMode mode,double lamda){
+	public SGDOptimizer(float lr,Optimizer.GMode mode,float lamda){
 		super(lr,mode,lamda);
 	}
 
@@ -27,8 +27,8 @@ public class SGDOptimizer extends Optimizer {
 		for(int i=0;i<params.size();i++){
 			Blob param = params.get(i);
 			Blob grad = gradient.get(i);
-			double[] paramData = param.getData();
-			double[] gradData = grad.getData();
+			float[] paramData = param.getData();
+			float[] gradData = grad.getData();
 			assert param.getSize()==grad.getSize():"param data size not equal gradient data size";
 			for(int j=0;j<param.getSize();j++){
 				paramData[j] -= lr*gradData[j];
@@ -42,13 +42,13 @@ public class SGDOptimizer extends Optimizer {
 		for(int i=0;i<params.size();i++){
 			Blob param = params.get(i);
 			Blob grad = gradient.get(i);
-			double[] paramData = param.getData();
-			double[] gradData = grad.getData();
+			float[] paramData = param.getData();
+			float[] gradData = grad.getData();
 			assert param.getSize()==grad.getSize():"param data size not equal gradient data size";
 			if(mode==GMode.L2) {
 				for(int j=0;j<param.getSize();j++){
 					//Ìí¼Ól2Ë¥¼õ
-					paramData[j] = (1.0-lr*lamda)*paramData[j]  - lr*gradData[j];
+					paramData[j] = (1.0f-lr*lamda)*paramData[j]  - lr*gradData[j];
 				}
 			}else if(mode==GMode.L1){
 				for(int j=0;j<param.getSize();j++){
