+++ src/test/cifar10/Cifar10Network.java
@@ -88,7 +88,7 @@ public class Cifar10Network {
 		network.setBatch(100);
 		network.setLoss(new LogLikeHoodLoss());
 		//network.setLoss(new CrossEntropyLoss());
-		optimizer = new SGDOptimizer(0.01,Optimizer.GMode.L2,0.0001);
+		optimizer = new SGDOptimizer(0.01f,Optimizer.GMode.L2,0.0001f);
 		network.setOptimizer(optimizer);
 		
 		//buildFcNetwork();
@@ -101,14 +101,14 @@ public class Cifar10Network {
 		Blob input = new Blob(batch,channel,height,width);
 		Blob label = new Blob(batch,network.getDatas().get(network.getDatas().size()-1).get3DSize(),1,1);
 		label.fillValue(0);
-		double[] blobData = input.getData();
-		double[] labelData = label.getData();
+		float[] blobData = input.getData();
+		float[] labelData = label.getData();
 		for(int i=start;i<(batch+start);i++){
 			DigitImage img = imageList.get(i);
 			byte[] imgData = img.imageData;
 			assert img.imageData.length== input.get3DSize():"buildBlobByImageList -- blob size error";
 			for(int j=0;j<imgData.length;j++){
-				blobData[(i-start)*input.get3DSize()+j] = (imgData[j]&0xff)/128.0-1;//normalize and centerlize(-1,1)
+				blobData[(i-start)*input.get3DSize()+j] = (imgData[j]&0xff)/128.0f-1;//normalize and centerlize(-1,1)
 			}
 			int labelValue = img.label;
 			for(int j=0;j<label.get3DSize();j++){
@@ -135,7 +135,7 @@ public class Cifar10Network {
 		return maxIndex;
 	}
 	
-	private int[] getBatchOutputLabel(double[] data){
+	private int[] getBatchOutputLabel(float[] data){
 		int[] outLabels = new int[network.getDatas().get(network.getDatas().size()-1).getNumbers()];
 		int outDataSize = network.getDatas().get(network.getDatas().size()-1).get3DSize();
 		for(int n=0;n<outLabels.length;n++){
@@ -171,20 +171,19 @@ public class Cifar10Network {
 	public void train(List<DigitImage> trainLists,int epoes,List<DigitImage> testLists){
 		System.out.println("training...... please wait for a moment!");
 		int batch = network.getBatch();
-		double loclaLr = optimizer.getLr();
-		double lossValue = 0.0;
+		float loclaLr = optimizer.getLr();
+		float lossValue = 0.0f;
 		for(int e=0;e<epoes;e++){
 			Collections.shuffle(trainLists);
 			long start = System.currentTimeMillis();
 			for(int i=0;i<=trainLists.size()-batch;i+=batch){
 				List<Blob> inputAndLabel = buildBlobByImageList(trainLists,i,batch,3,32,32);
-				double tmpLoss = network.train(inputAndLabel.get(0), inputAndLabel.get(1));
+				float tmpLoss = network.train(inputAndLabel.get(0), inputAndLabel.get(1));
 				lossValue = (lossValue+tmpLoss)/2;
 				if(i%1000==0) {
 					System.out.print(".");
 				}
 			}
-			//每个epoe做一次测试
 			//每个epoe做一次测试
 			System.out.println();
 			System.out.println("training...... epoe: "+e+" lossValue: "+lossValue
@@ -192,8 +191,8 @@ public class Cifar10Network {
 		
 			test(testLists);
 			
-			if(loclaLr>0.001){
-				loclaLr*=0.8;
+			if(loclaLr>0.001f){
+				loclaLr*=0.8f;
 				optimizer.setLr(loclaLr);
 			}
 		}
